+++ cuneiform_src/Kern/hrstr/layout.h
@@ -201,14 +201,14 @@ void RotatePageToIdeal (void);
 
 # define IDEAL_XY(x, y)   \
          {\
-             x = (INT) (x + (LONG) y * nIncline / INCLINE_FACTOR);\
-             y = (INT) (y - (LONG) x * nIncline / INCLINE_FACTOR);\
+             x = (INT) (x + (int) y * nIncline / INCLINE_FACTOR);\
+             y = (INT) (y - (int) x * nIncline / INCLINE_FACTOR);\
          }
 
 # define REAL_XY(x, y)   \
          {\
-             y = (INT) (y + (LONG) x * nIncline / INCLINE_FACTOR);\
-             x = (INT) (x - (LONG) y * nIncline / INCLINE_FACTOR);\
+             y = (INT) (y + (int) x * nIncline / INCLINE_FACTOR);\
+             x = (INT) (x - (int) y * nIncline / INCLINE_FACTOR);\
          }
 
 /****************************************************************************
